requirements.txt
aiosql==9.0
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
psycopg2==2.9.7
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.4
